chore: rename ui directory to scene
