+++ src/index.ts
@@ -3,4 +3,4 @@ import * as PixiSound from '@pixi/sound';
 import * as WebFontLoader from 'webfontloader';
 
 export { Pixi, PixiSound, WebFontLoader };
-export * from './ui';
+export * from './scene';
